planning-platform/backend/requirements.txt
# 김현우내과 건강검진 백엔드 의존성

# FastAPI 관련
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# 데이터베이스 관련
asyncpg==0.29.0
databases[postgresql]==0.8.0
sqlalchemy==1.4.53
alembic==1.13.1

# 인증 및 보안
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-decouple==3.8

# 세션 관리
redis==5.0.1
aioredis==2.0.1

# 유틸리티
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
httpx==0.25.2

# 개발 도구
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0
mypy==1.7.1